+++ internal/leadership/elector_test.go
@@ -230,13 +230,13 @@ func testElector[TElectorBundle any](
 		// The leadership maintenance loop also listens on the leadership
 		// notification channel. Take advantage of that to cause an
 		// immediate reelect attempt with no sleep.
-		elector.leadershipNotificationChan <- struct{}{}
+		elector.leaderResignedChan <- struct{}{}
 		elector.testSignals.MaintainedLeadership.WaitOrTimeout()
 
-		elector.leadershipNotificationChan <- struct{}{}
+		elector.leaderResignedChan <- struct{}{}
 		elector.testSignals.MaintainedLeadership.WaitOrTimeout()
 
-		elector.leadershipNotificationChan <- struct{}{}
+		elector.leaderResignedChan <- struct{}{}
 		elector.testSignals.MaintainedLeadership.WaitOrTimeout()
 
 		elector.Stop()
@@ -269,7 +269,7 @@ func testElector[TElectorBundle any](
 			Schema:   elector.config.Schema,
 		})
 
-		elector.leadershipNotificationChan <- struct{}{}
+		elector.leaderResignedChan <- struct{}{}
 		elector.testSignals.LostLeadership.WaitOrTimeout()
 
 		// Wait for the elector to try and fail to gain leadership so we
@@ -317,7 +317,7 @@ func testElector[TElectorBundle any](
 			// Cheat if we're in poll only by notifying leadership channel to
 			// wake the elector from sleep.
 			if elector2.notifier == nil {
-				elector2.leadershipNotificationChan <- struct{}{}
+				elector2.leaderResignedChan <- struct{}{}
 			}
 
 			t.Logf("Waiting for %s to gain leadership", elector2.config.ClientID)
@@ -474,9 +474,10 @@ func TestElectorHandleLeadershipNotification(t *testing.T) {
 			&Config{ClientID: "test_client_id"},
 		)
 
-		// This channel is normally only initialized on start, so we need to
+		// These channels are normally only initialized on start, so we need to
 		// create it manually here.
-		elector.leadershipNotificationChan = make(chan struct{}, 1)
+		elector.forceResignChan = make(chan struct{}, 1)
+		elector.leaderResignedChan = make(chan struct{}, 1)
 
 		return elector, &testBundle{}
 	}
@@ -489,11 +490,11 @@ func TestElectorHandleLeadershipNotification(t *testing.T) {
 		return data
 	}
 
-	validLeadershipChange := func() *dbLeadershipNotification {
+	validLeadershipChange := func() *DBNotification {
 		t.Helper()
 
-		return &dbLeadershipNotification{
-			Action:   "resigned",
+		return &DBNotification{
+			Action:   DBNotificationKindResigned,
 			LeaderID: "other-client-id",
 		}
 	}
@@ -505,7 +506,7 @@ func TestElectorHandleLeadershipNotification(t *testing.T) {
 
 		elector.handleLeadershipNotification(ctx, notifier.NotificationTopicLeadership, string(mustMarshalJSON(t, validLeadershipChange())))
 
-		riversharedtest.WaitOrTimeout(t, elector.leadershipNotificationChan)
+		riversharedtest.WaitOrTimeout(t, elector.leaderResignedChan)
 	})
 
 	t.Run("StopsOnContextDone", func(t *testing.T) {
@@ -518,7 +519,7 @@ func TestElectorHandleLeadershipNotification(t *testing.T) {
 
 		elector.handleLeadershipNotification(ctx, notifier.NotificationTopicLeadership, string(mustMarshalJSON(t, validLeadershipChange())))
 
-		require.Empty(t, elector.leadershipNotificationChan)
+		require.Empty(t, elector.leaderResignedChan)
 	})
 
 	t.Run("IgnoresNonResignedAction", func(t *testing.T) {
@@ -531,7 +532,7 @@ func TestElectorHandleLeadershipNotification(t *testing.T) {
 
 		elector.handleLeadershipNotification(ctx, notifier.NotificationTopicLeadership, string(mustMarshalJSON(t, change)))
 
-		require.Empty(t, elector.leadershipNotificationChan)
+		require.Empty(t, elector.leaderResignedChan)
 	})
 
 	t.Run("IgnoresSameClientID", func(t *testing.T) {
@@ -544,6 +545,6 @@ func TestElectorHandleLeadershipNotification(t *testing.T) {
 
 		elector.handleLeadershipNotification(ctx, notifier.NotificationTopicLeadership, string(mustMarshalJSON(t, change)))
 
-		require.Empty(t, elector.leadershipNotificationChan)
+		require.Empty(t, elector.leaderResignedChan)
 	})
 }